initial verion of resource discovery.
This is a rewrite and clean up of an old much used system. The concepts are solid but need to do some real beta testing
prior to marking this 1.0.  This is very useful stuff and needs to make it's way out.  For that to happen extensive docs need to be created yet.